Make Travis publish examples to github.io
This commit makes Travis publish the standalone examples to https://camptocamp.github.io/ngeo/master/. This is done by running the gh-pages-from-travis Makefile target in an after_success hook, only for the master branch, and when the build is not the run as a result of a pull request.

The secure variable in .travis.yml is created using this command:

$ travis encrypt -r camptocamp/ngeo 'GH_TOKEN=<token>'

where <token> is a "GitHub Personal access token" created on https://github.com/settings/applications.