Select print style type based on geometry type
With this commint the vector style type used in the print spec is selected based on the geometry type, instead of being selected based on the properties found in the ol.style.Style object. Relying on the properties found in the ol.style.Style object is fragile, as a style object may be used for drawing different types of geometries.