Profile: allow changing scales from application code
We used to provide two options: yLowerBound and ratioXYRule to allow
applications to change the Y domain of the profile. This allowed enforcing
rules like: min y value > 0 or a specific aspect ratio.

It was not working so well though since the ratioXYRule signature was too
restricted (only taking a xDomainMax) which could lead to ugly results.

In this commit, these options are replaced by a more powerful option.