Fix menu animation on Safari
In a sidebar, when clicking on a link in the main menu (e.g. "Background"),
the title bar turned white during the animation on Safari.
The problem is that the newly inserted element did not have the right position
after insertion on Safari. The "fix" forces Safari to calculate the position.