Fix layertree parent state
When the first child is checked and there exist another child which
is not checked, the parent was incorrectly checked instead of being
indeterminate.

To avoid this, the parent state is always refreshed.

The root cause is that state management is done too early, when the
layertree controller hierarchy is instanciated.

A more involved fix would be to:
- stop initializing layertrees in their constructors;
- have layertrees register a ready listener on the root layertree;
- have the root layertree triggers an onReady event in its postLink
  function.