Generate profile csv on client-side
When starting a CSV export for a profile, a request was made to export the
profile data as CSV. And then a download is started on the client-side.
The problem is that Safari requires that the download is started within 1s
after a user interaction ('trusted event', e.g. clicking the export button).
But by doing the request, this requirement is not fulfilled. To avoid the
request, this PR generates the CSV on the client-side by using the already
available data.
See also: https://github.com/eligrey/FileSaver.js/issues/12