Remove more dependencies
Remove core-js:
It looks to be a bunch of pollyfill.
I would like to remove it, and then ev. patch the issues with
dedicated well-know pollyfill

regenerator-runtime:
Was imported in src/controllers/bootstrap.js, but probably not used
anymore.

Cheerio
Is probably not worth to fix it as subdependencies

resize-observer-polyfill
Should not be necessary anymore in 2024