Really fix edit panel activation
Follow-up of PR #9909.

Changing the panel name in the tool button is a breaking change since it requires to change the slot name in the main HTML document, so change the panel name in the listener instead.

Additionnally, there still was a mismatch in the name of the AbstractDesktopController class field and the tool active property name (editFeatureActive)